refactored states, cities and address on actions
