+++ src/app/(auth)/social-assistant/organizations/[organizationId]/subjects/[subjectId]/edit/_actions.ts
@@ -25,6 +25,17 @@ export async function updateSubjectAction(
   prevState: any,
   formData: FormData,
 ): Promise<ApiResponse> {
+  const address = {
+    street: formData.get("street"),
+    number: formData.get("number"),
+    district: formData.get("district"),
+    complement: formData.get("complement"),
+    state_id: formData.get("state_id"),
+    city_id: formData.get("city_id"),
+  };
+
+  const addressHasValue = Object.values(address).filter((a) => a != "" && a).length;
+
   const response = await api({
     input: `/${Roles.SOCIAL_ASSISTANT}/subjects/${subjectId}`,
     init: {
@@ -42,16 +53,7 @@ export async function updateSubjectAction(
         relative_relation_type: formData.get("relative_relation_type"),
         relative_name: formData.get("relative_name"),
         relative_phone: formData.get("relative_phone"),
-        addresses: [
-          {
-            street: formData.get("street"),
-            number: formData.get("number"),
-            district: formData.get("district"),
-            complement: formData.get("complement"),
-            state_id: formData.get("state_id"),
-            city_id: formData.get("city_id"),
-          },
-        ],
+        addresses: addressHasValue ? [address] : null,
       }),
     },
   });
