+++ src/app/(auth)/social-assistant/organizations/[organizationId]/subjects/[subjectId]/edit/_form.tsx
@@ -169,7 +169,6 @@ export function UpdateSubjectForm({ subject, skinColors, states, cities }: Updat
         name="state_id"
         label={t("labels.state")}
         placeholder={t("labels.state")}
-        withAsterisk
         data={states.map((state) => ({ label: state.name, value: state.id.toString() }))}
         queryName="state"
         errors={state?.errors?.["addresses.0.state_id"]}
@@ -180,7 +179,6 @@ export function UpdateSubjectForm({ subject, skinColors, states, cities }: Updat
         name="city_id"
         label={t("labels.city")}
         placeholder={t("labels.city")}
-        withAsterisk
         data={cities.map((city) => ({ label: city.name, value: city.id.toString() }))}
         disabled={cities.length == 0}
         errors={state?.errors?.["addresses.0.city_id"]}
@@ -191,7 +189,6 @@ export function UpdateSubjectForm({ subject, skinColors, states, cities }: Updat
         name="street"
         label={t("labels.street")}
         placeholder={t("labels.street")}
-        withAsterisk
         errors={state?.errors?.["addresses.0.street"]}
         defaultValue={address?.street}
       />
@@ -200,7 +197,6 @@ export function UpdateSubjectForm({ subject, skinColors, states, cities }: Updat
         name="number"
         label={t("labels.number")}
         placeholder={t("labels.number")}
-        withAsterisk
         errors={state?.errors?.["addresses.0.number"]}
         defaultValue={address?.number}
       />
