+++ src/app/(auth)/social-assistant/organizations/[organizationId]/subjects/_actions.ts
@@ -1,7 +1,6 @@
 "use server";
 
 import api from "@/api";
-import { FlashMessageTypes } from "@/enums/FlashMessageTypes";
 import { Roles } from "@/enums/Roles";
 import { ApiResponse } from "@/schemas";
 import { revalidateTag } from "next/cache";
@@ -13,29 +12,6 @@ export async function fetchFormInfo(): Promise<ApiResponse> {
       method: "GET",
     },
   });
-
-  const json = await response.json();
-
-  return ApiResponse.parse(json);
-}
-
-export async function fetchCitiesByState(stateId?: number): Promise<ApiResponse> {
-  const response = await api({
-    input: `/${Roles.SOCIAL_ASSISTANT}/subjects/get/cities-by-state-to-select/${stateId}`,
-    init: {
-      method: "GET",
-    },
-  });
-
-  if (!response.ok) {
-    return ApiResponse.parse({
-      type: FlashMessageTypes.SUCCESS,
-      message: "",
-      data: {
-        cities: [],
-      },
-    });
-  }
 
   const json = await response.json();
