+++ src/app/(auth)/social-assistant/organizations/[organizationId]/subjects/create/_actions.ts
@@ -11,6 +11,17 @@ export async function createSubjectAction(
   prevState: any,
   formData: FormData,
 ): Promise<ApiResponse> {
+  const address = {
+    street: formData.get("street"),
+    number: formData.get("number"),
+    district: formData.get("district"),
+    complement: formData.get("complement"),
+    state_id: formData.get("state_id"),
+    city_id: formData.get("city_id"),
+  };
+
+  const addressHasValue = Object.values(address).filter((a) => a != "" && a).length;
+
   const response = await api({
     input: `/${Roles.SOCIAL_ASSISTANT}/subjects/${organizationId}`,
     init: {
@@ -28,16 +39,7 @@ export async function createSubjectAction(
         relative_relation_type: formData.get("relative_relation_type"),
         relative_name: formData.get("relative_name"),
         relative_phone: formData.get("relative_phone"),
-        addresses: [
-          {
-            street: formData.get("street"),
-            number: formData.get("number"),
-            district: formData.get("district"),
-            complement: formData.get("complement"),
-            state_id: formData.get("state_id"),
-            city_id: formData.get("city_id"),
-          },
-        ],
+        addresses: addressHasValue ? [address] : null,
       }),
     },
   });
