+++ src/app/(auth)/social-assistant/organizations/[organizationId]/subjects/create/page.tsx
@@ -1,10 +1,11 @@
+import { fetchCitiesByState, fetchStates } from "@/app/(auth)/_actions";
 import { Paper } from "@/components/containers";
 import { Heading } from "@/components/ui";
 import { t } from "@/lang";
 import { State } from "@/schemas";
 import { City } from "@/schemas/City";
 import { SkinColor } from "@/schemas/SkinColor";
-import { fetchCitiesByState, fetchFormInfo } from "../_actions";
+import { fetchFormInfo } from "../_actions";
 import { CreateSubjectForm } from "./_form";
 
 interface PageProps {
@@ -15,11 +16,12 @@ interface PageProps {
 
 export default async function page({ searchParams }: PageProps) {
   const { data: formInfo } = await fetchFormInfo();
+  const { data: statesData } = await fetchStates();
   const { data: citiesData } = await fetchCitiesByState(searchParams.state);
 
   const skinColors = SkinColor.array().parse(formInfo.skinColors);
-  const states = State.array().parse(formInfo.states);
-  const cities = City.array().parse(citiesData.cities);
+  const states = State.array().parse(statesData);
+  const cities = City.array().parse(citiesData);
 
   return (
     <>
